+++ recaf-launcher/build.gradle
@@ -7,6 +7,6 @@ dependencies {
 
 jar {
     manifest {
-        attributes('Main-Class': 'dev.xdark.launcher.Launcher')
+        attributes('Main-Class': 'dev.xdark.recaf.Launcher')
     }
 }
