+++ recaf-launcher/src/main/java/dev/xdark/recaf/InstantTypeAdapter.java
@@ -1,4 +1,4 @@
-package dev.xdark.launcher;
+package dev.xdark.recaf;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
